+++ packages/stream_chat/lib/src/event_type.dart
@@ -170,4 +170,11 @@ class EventType {
 
   /// Event sent when a message reminder is due.
   static const String notificationReminderDue = 'notification.reminder_due';
+
+  /// Local event sent when push notification preference is updated.
+  static const String pushPreferenceUpdated = 'push_preference.updated';
+
+  /// Local event sent when channel push notification preference is updated.
+  static const String channelPushPreferenceUpdated =
+      'channel.push_preference.updated';
 }